requirements.txt
Django==3.2.6
djangorestframework==3.12.4
Pillow==8.3.1
django-cors-headers==3.8.0